+++ src/constants/char_constants.py
@@ -1,0 +1,7 @@
+class TextSymbols:
+    square_f: str = "\u25a3"
+    square_e: str = "\u25a1"
+    check_mark: str = "\u2713"
+    left_arrow: str = "\u2190"
+    x_symbol: str = "\u0078"
+    music_symbol: str = "\u266a"
